+++ pytest_selenium/drivers/chrome.py
@@ -2,9 +2,19 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
+import pytest
+from selenium.webdriver.chrome.options import Options
 
-def driver_kwargs(capabilities, driver_path, **kwargs):
-    kwargs = {'desired_capabilities': capabilities}
+
+def driver_kwargs(capabilities, driver_path, chrome_options, **kwargs):
+    kwargs = {
+        'desired_capabilities': capabilities,
+        'chrome_options': chrome_options}
     if driver_path is not None:
         kwargs['executable_path'] = driver_path
     return kwargs
+
+
+@pytest.fixture
+def chrome_options():
+    return Options()